Use polib for wrapping
This removes the dependency on msgcat, improving the experience for
Windows people, and eliminating potential wrapping discrepancies caused
by different versions of msgcat.

It also obviates creating a temporary file.

Known issue: polib's wrapping is different from msgcat's.  It doesn't
always put the space at the start of the line. This also means that
files wrapped with a previous version of powrap will likely be
completely reformatted.

Fixes #76, #82, #104